tech: Move generate funcs out of MockServer
